% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower.R
\name{gower_dist}
\alias{gower_dist}
\title{Gower's distace}
\usage{
gower_dist(x, y, ranges)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}
}
\description{
Recycling over the data.frame with the least records.
}

